add monthly income-expense report
My monthly income and expenses vary a bit during the year. I usually generate this report manually to know which month has been expensive and check if the income looks reasonable for each month. This level of detail is a lot more helpful for me to see how I am doing in the current year.